# OneMoreTaskTracker.Api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OneMoreTaskTracker Plan API",
    "version": "v2",
    "description": "REST surface emitted by OneMoreTaskTracker.Api for the feature planning (Gantt) experience. Backed by gRPC into OneMoreTaskTracker.Features / .Tasks / .Users. v2 replaces the flat 5-stage taxonomy with a structured gate(3) + track(2) + phase(4) + sub-stage(1..6) taxonomy. Hand-rolled per memory `project_openapi_hand_rolled`; drift = contract auto-fail."
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local dev" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/auth/health": {
      "get": {
        "summary": "Liveness probe — returns {status:'ok'}.",
        "description": "Anonymous, no upstream calls. Used by deployment + GAN harness (BE_HEALTH_URL).",
        "tags": ["auth"],
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status"],
                  "properties": { "status": { "type": "string", "enum": ["ok"] } }
                }
              }
            }
          }
        }
      }
    },
    "/api/plan/features": {
      "get": {
        "summary": "List features visible to the caller.",
        "description": "Optional `windowStart` / `windowEnd` query params (ISO yyyy-MM-dd) constrain results to features whose plan overlaps the window. Bad date format or `windowStart > windowEnd` returns 400 `Invalid request data` and the upstream gRPC service is NOT called.",
        "tags": ["plan"],
        "parameters": [
          { "name": "scope", "in": "query", "required": false, "schema": { "type": "string", "enum": ["mine", "all"] } },
          { "name": "state", "in": "query", "required": false, "schema": { "$ref": "#/components/schemas/FeatureState" } },
          { "name": "windowStart", "in": "query", "required": false, "description": "ISO yyyy-MM-dd. Inclusive on plannedEnd.", "schema": { "type": "string", "format": "date" } },
          { "name": "windowEnd", "in": "query", "required": false, "description": "ISO yyyy-MM-dd. Exclusive on plannedStart.", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/FeatureSummary" } }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "summary": "Create a new feature (Manager only).",
        "description": "Creates the feature and atomically materializes the v2 taxonomy: 3 gates in `waiting` (spec, backend.prep-gate, frontend.prep-gate) + 8 phases each with exactly 1 unowned sub-stage. `managerUserId` is derived from the caller's JWT and MUST NOT be sent in the body.",
        "tags": ["plan"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateFeaturePayload" } }
          }
        },
        "responses": {
          "200": {
            "description": "Created",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/FeatureSummary" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/plan/features/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int32" } },
        { "name": "If-Match", "in": "header", "required": false, "description": "Optional optimistic-concurrency token; current feature.version from the last read. Forwarded as `expectedVersion` on the sparse path.", "schema": { "type": "string" } }
      ],
      "get": {
        "summary": "Get a single feature with attached tasks, lead, mini-team, and full taxonomy.",
        "tags": ["plan"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/FeatureDetail" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "summary": "Update a feature — sparse PATCH (Manager only).",
        "description": "Consolidated PATCH endpoint: send only the fields you want to change. Sparse fields supported: `title`, `description`, `leadUserId`, `expectedVersion`. Omitted fields are unchanged. Per-gate edits live at `/api/plan/features/{id}/gates/{gateKey}`. Per-sub-stage edits live at `/api/plan/features/{id}/sub-stages/{subStageId}`. Optimistic concurrency: include `If-Match` header with the current `feature.version`, or `expectedVersion` in the body. Body wins if both are supplied.",
        "tags": ["plan"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateFeaturePayload" } }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/FeatureSummary" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "502": { "$ref": "#/components/responses/BadGateway" }
        }
      }
    },
    "/api/plan/features/{id}/gates/{gateKey}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int32" } },
        { "name": "gateKey", "in": "path", "required": true, "description": "One of `spec`, `backend.prep-gate`, `frontend.prep-gate`.", "schema": { "type": "string", "enum": ["spec", "backend.prep-gate", "frontend.prep-gate"] } },
        { "name": "If-Match", "in": "header", "required": false, "description": "Optional gate-scoped optimistic-concurrency token (taxonomy.gates[].gateVersion). Forwarded as `expectedVersion`.", "schema": { "type": "string" } }
      ],
      "patch": {
        "summary": "Flip a gate's approval status — sparse PATCH (Manager only).",
        "description": "Sparse PATCH; only fields the caller wants to change are sent. Server stamps `approverUserId` + `approvedAtUtc`. `gateStatus = approved` clears any prior `rejectionReason`. `gateStatus = rejected` requires a non-empty `rejectionReason` (≤ 500 chars). `gateStatus = waiting` clears approver, timestamp, reason and stamps a new `requestedAtUtc`. Optimistic concurrency: `If-Match` header carries `gate.version` or `expectedVersion` in body; body wins.",
        "tags": ["plan"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PatchFeatureGatePayload" } }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PatchFeatureGateResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "502": { "$ref": "#/components/responses/BadGateway" }
        }
      }
    },
    "/api/plan/features/{id}/sub-stages/{subStageId}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int32" } },
        { "name": "subStageId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int32" } },
        { "name": "If-Match", "in": "header", "required": false, "description": "Optional sub-stage-scoped optimistic-concurrency token (taxonomy.tracks[].phases[].subStages[].subStageVersion). Forwarded as `expectedVersion`.", "schema": { "type": "string" } }
      ],
      "patch": {
        "summary": "Update a sub-stage — sparse PATCH (Manager only).",
        "description": "Sparse PATCH of one sub-stage's `ownerUserId` / `plannedStart` / `plannedEnd`. Send only the fields you want to change. Validators: `plannedStart ≤ plannedEnd` if both present, dates ISO yyyy-MM-dd in years 2000..2100. Optimistic concurrency: `If-Match` header carries `subStage.version` or `expectedVersion` in body; body wins.",
        "tags": ["plan"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PatchFeatureSubStagePayload" } }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SubStageMutationResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "502": { "$ref": "#/components/responses/BadGateway" }
        }
      },
      "delete": {
        "summary": "Delete a sub-stage (Manager only).",
        "description": "Removes one owner sub-stage. The phase must retain ≥ 1 sub-stage; deleting the last sub-stage of a phase is rejected 409 with `conflict.kind = 'subStageCap'`. `ethalon-testing` and `live-release` reject deletes outright. `If-Match` optional.",
        "tags": ["plan"],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SubStageMutationResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/plan/features/{id}/phases/{track}/{phase}/sub-stages": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int32" } },
        { "name": "track", "in": "path", "required": true, "schema": { "type": "string", "enum": ["backend", "frontend"] } },
        { "name": "phase", "in": "path", "required": true, "schema": { "type": "string", "enum": ["development", "stand-testing"] }, "description": "Only multi-owner phases accept appends; ethalon-testing / live-release are capped at 1." }
      ],
      "post": {
        "summary": "Append a sub-stage to a multi-owner phase (Manager only).",
        "description": "Appends a new owner sub-stage at the next ordinal. Server-fills `plannedStart` from previous sibling's `plannedEnd` and `plannedEnd` to `plannedStart + 1d` if either is null. Hard cap = 6 sub-stages per phase per track; exceeding returns 409 with `conflict.kind = 'subStageCap'`.",
        "tags": ["plan"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AppendFeatureSubStagePayload" } }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SubStageMutationResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/plan/features/{id}/tasks/{jiraId}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int32" } },
        { "name": "jiraId", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "post": {
        "summary": "Attach a task (by Jira id) to a feature (Manager only).",
        "tags": ["plan"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/FeatureSummary" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "summary": "Detach a task from a feature, reassigning it to another feature (Manager only).",
        "tags": ["plan"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/DetachTaskBody" } }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/FeatureSummary" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request data.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Authentication required.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "Permission denied.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "UnprocessableEntity": {
        "description": "Validation error.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConflictError" } } }
      },
      "BadGateway": {
        "description": "Upstream service temporarily unavailable.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Conflict": {
        "description": "Version mismatch (concurrent edit detected) or sub-stage cap reached. Body extends to `{ error, conflict: { kind: 'version' | 'subStageCap' | 'subStageOverlap', currentVersion?, track?, phase?, cap? } }` so the FE can render the right CTA.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConflictError" } } }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      },
      "ConflictError": {
        "type": "object",
        "description": "Extended error envelope emitted on 409/422 when the handler reports a structured conflict. Clients unaware of `conflict` fall back to the bare `error` string.",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Public-facing message such as 'Updated by someone else' or 'Sub-stage cap reached'." },
          "conflict": {
            "type": "object",
            "description": "Structured conflict context. Absent when the upstream did not supply one.",
            "required": ["kind"],
            "properties": {
              "kind": { "type": "string", "enum": ["version", "subStageCap", "subStageOverlap"] },
              "currentVersion": { "type": "integer", "format": "int32", "description": "Server's current version (kind='version')." },
              "track":          { "type": "string", "enum": ["Backend", "Frontend"], "description": "Track whose phase tripped a cap (kind='subStageCap'/'subStageOverlap')." },
              "phase":          { "type": "string", "description": "Phase whose cap tripped (kind='subStageCap'/'subStageOverlap')." },
              "cap":            { "type": "integer", "format": "int32", "description": "Hard cap value (kind='subStageCap')." }
            }
          }
        }
      },
      "FeatureState": {
        "type": "string",
        "enum": ["CsApproving", "Development", "Testing", "EthalonTesting", "LiveRelease"]
      },
      "GateKind": {
        "type": "string",
        "enum": ["spec", "cs", "sr"]
      },
      "GateStatus": {
        "type": "string",
        "enum": ["waiting", "approved", "rejected"]
      },
      "Track": {
        "type": "string",
        "enum": ["backend", "frontend"]
      },
      "PhaseKind": {
        "type": "string",
        "enum": ["development", "stand-testing", "ethalon-testing", "live-release"]
      },
      "FeatureGate": {
        "type": "object",
        "required": ["id", "gateKey", "kind", "status", "version"],
        "properties": {
          "id":              { "type": "integer", "format": "int32" },
          "gateKey":         { "type": "string", "enum": ["spec", "backend.prep-gate", "frontend.prep-gate"] },
          "kind":            { "$ref": "#/components/schemas/GateKind" },
          "track":           { "nullable": true, "allOf": [ { "$ref": "#/components/schemas/Track" } ], "description": "null on the spec gate; set on track-scoped gates." },
          "status":          { "$ref": "#/components/schemas/GateStatus" },
          "approverUserId":  { "type": "integer", "format": "int32", "nullable": true, "description": "Bare cross-service id (Users service); resolution lives in `miniTeam` on detail responses." },
          "approvedAtUtc":   { "type": "string", "format": "date-time", "nullable": true },
          "requestedAtUtc":  { "type": "string", "format": "date-time", "nullable": true, "description": "Set on creation and on flips back to waiting." },
          "rejectionReason": { "type": "string", "nullable": true, "maxLength": 500 },
          "version":         { "type": "integer", "format": "int32", "description": "Per-gate optimistic-concurrency token." }
        }
      },
      "FeatureSubStage": {
        "type": "object",
        "required": ["id", "track", "phase", "ordinal", "version"],
        "properties": {
          "id":           { "type": "integer", "format": "int32" },
          "track":        { "$ref": "#/components/schemas/Track" },
          "phase":        { "$ref": "#/components/schemas/PhaseKind" },
          "ordinal":      { "type": "integer", "format": "int32", "minimum": 1, "description": "1-based position within the phase, recomputed on append/delete." },
          "ownerUserId":  { "type": "integer", "format": "int32", "nullable": true },
          "plannedStart": { "type": "string", "format": "date", "nullable": true },
          "plannedEnd":   { "type": "string", "format": "date", "nullable": true },
          "version":      { "type": "integer", "format": "int32", "description": "Per-sub-stage optimistic-concurrency token." }
        }
      },
      "FeaturePhaseTaxonomy": {
        "type": "object",
        "required": ["phase", "multiOwner", "cap", "subStages"],
        "properties": {
          "phase":      { "$ref": "#/components/schemas/PhaseKind" },
          "multiOwner": { "type": "boolean", "description": "true for development/stand-testing; false for ethalon-testing/live-release." },
          "cap":        { "type": "integer", "format": "int32", "description": "Hard cap on sub-stages — 6 for multi-owner phases, 1 otherwise." },
          "subStages":  { "type": "array", "items": { "$ref": "#/components/schemas/FeatureSubStage" } }
        }
      },
      "FeatureTrackTaxonomy": {
        "type": "object",
        "required": ["track", "phases"],
        "properties": {
          "track":  { "$ref": "#/components/schemas/Track" },
          "phases": { "type": "array", "minItems": 4, "maxItems": 4, "items": { "$ref": "#/components/schemas/FeaturePhaseTaxonomy" }, "description": "Always length 4 in canonical order: development, stand-testing, ethalon-testing, live-release." }
        }
      },
      "FeatureTaxonomy": {
        "type": "object",
        "required": ["gates", "tracks"],
        "properties": {
          "gates":  { "type": "array", "minItems": 3, "maxItems": 3, "items": { "$ref": "#/components/schemas/FeatureGate" }, "description": "Always length 3 in canonical order: spec, backend.prep-gate, frontend.prep-gate." },
          "tracks": { "type": "array", "minItems": 2, "maxItems": 2, "items": { "$ref": "#/components/schemas/FeatureTrackTaxonomy" }, "description": "Always length 2: backend, frontend." }
        }
      },
      "MiniTeamMember": {
        "type": "object",
        "required": ["userId", "email", "displayName", "role"],
        "properties": {
          "userId":      { "type": "integer", "format": "int32" },
          "email":       { "type": "string" },
          "displayName": { "type": "string" },
          "role":        { "type": "string" }
        }
      },
      "AttachedTask": {
        "type": "object",
        "required": ["id", "jiraId", "state", "userId"],
        "properties": {
          "id":     { "type": "integer", "format": "int32" },
          "jiraId": { "type": "string" },
          "state":  { "type": "string" },
          "userId": { "type": "integer", "format": "int32" }
        }
      },
      "FeatureSummary": {
        "type": "object",
        "required": ["id","title","description","state","plannedStart","plannedEnd","leadUserId","managerUserId","taskCount","taskIds","taxonomy"],
        "properties": {
          "id":            { "type": "integer", "format": "int32" },
          "title":         { "type": "string", "maxLength": 200 },
          "description":   { "type": "string", "nullable": true, "maxLength": 4000 },
          "state":         { "$ref": "#/components/schemas/FeatureState" },
          "plannedStart":  { "type": "string", "format": "date", "nullable": true, "description": "Manually editable feature-scoped date." },
          "plannedEnd":    { "type": "string", "format": "date", "nullable": true, "description": "Manually editable feature-scoped date." },
          "leadUserId":    { "type": "integer", "format": "int32" },
          "managerUserId": { "type": "integer", "format": "int32" },
          "taskCount":     { "type": "integer", "format": "int32" },
          "taskIds":       { "type": "array", "items": { "type": "integer", "format": "int32" } },
          "taxonomy":      { "$ref": "#/components/schemas/FeatureTaxonomy" },
          "version":       { "type": "integer", "format": "int32", "description": "Monotonically-increasing feature-scoped optimistic-concurrency token; bumped by every feature edit, gate flip, or sub-stage mutation." }
        }
      },
      "FeatureDetail": {
        "type": "object",
        "required": ["feature","tasks","lead","miniTeam"],
        "properties": {
          "feature":  { "$ref": "#/components/schemas/FeatureSummary" },
          "tasks":    { "type": "array", "items": { "$ref": "#/components/schemas/AttachedTask" } },
          "lead":     { "$ref": "#/components/schemas/MiniTeamMember" },
          "miniTeam": { "type": "array", "items": { "$ref": "#/components/schemas/MiniTeamMember" }, "description": "Deduped union of task assignees, lead, every populated sub-stage owner, and every gate approver." }
        }
      },
      "CreateFeaturePayload": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title":        { "type": "string", "maxLength": 200 },
          "description":  { "type": "string", "nullable": true, "maxLength": 4000 },
          "leadUserId":   { "type": "integer", "format": "int32", "nullable": true },
          "plannedStart": { "type": "string", "format": "date", "nullable": true },
          "plannedEnd":   { "type": "string", "format": "date", "nullable": true }
        }
      },
      "UpdateFeaturePayload": {
        "type": "object",
        "description": "All fields optional — send only the fields you want to change. Sparse fields: `title`, `description`, `leadUserId`, `expectedVersion`.",
        "properties": {
          "title":           { "type": "string", "nullable": true, "maxLength": 200 },
          "description":     { "type": "string", "nullable": true, "maxLength": 4000 },
          "leadUserId":      { "type": "integer", "format": "int32", "nullable": true, "minimum": 1 },
          "expectedVersion": { "type": "integer", "format": "int32", "nullable": true, "description": "Optimistic-concurrency token. Body wins over `If-Match` header when both are supplied." }
        }
      },
      "PatchFeatureGatePayload": {
        "type": "object",
        "description": "All fields optional — send only the fields you want to change. Sparse fields: `status`, `rejectionReason`, `expectedVersion`.",
        "properties": {
          "status":          { "$ref": "#/components/schemas/GateStatus" },
          "rejectionReason": { "type": "string", "nullable": true, "maxLength": 500, "description": "Required only when status = 'rejected'." },
          "expectedVersion": { "type": "integer", "format": "int32", "nullable": true, "description": "Gate-scoped optimistic-concurrency token. Body wins over `If-Match` when both are supplied." }
        }
      },
      "PatchFeatureGateResponse": {
        "type": "object",
        "required": ["featureId", "featureVersion", "taxonomy"],
        "properties": {
          "featureId":      { "type": "integer", "format": "int32" },
          "featureVersion": { "type": "integer", "format": "int32" },
          "taxonomy":       { "$ref": "#/components/schemas/FeatureTaxonomy" }
        }
      },
      "PatchFeatureSubStagePayload": {
        "type": "object",
        "description": "All fields optional — send only the fields you want to change. Sparse fields: `ownerUserId`, `plannedStart`, `plannedEnd`, `expectedVersion`.",
        "properties": {
          "ownerUserId":     { "type": "integer", "format": "int32", "nullable": true, "minimum": 0 },
          "plannedStart":    { "type": "string",  "format": "date",  "nullable": true },
          "plannedEnd":      { "type": "string",  "format": "date",  "nullable": true },
          "expectedVersion": { "type": "integer", "format": "int32", "nullable": true, "description": "Sub-stage-scoped optimistic-concurrency token. Body wins over `If-Match`." }
        }
      },
      "AppendFeatureSubStagePayload": {
        "type": "object",
        "description": "All fields optional — server fills sensible defaults (planned dates from previous sibling).",
        "properties": {
          "ownerUserId":  { "type": "integer", "format": "int32", "nullable": true },
          "plannedStart": { "type": "string", "format": "date", "nullable": true },
          "plannedEnd":   { "type": "string", "format": "date", "nullable": true }
        }
      },
      "SubStageMutationResponse": {
        "type": "object",
        "description": "Returned by sub-stage append / patch / delete endpoints. `createdSubStageId` is set only on append.",
        "required": ["featureId", "featureVersion", "taxonomy"],
        "properties": {
          "featureId":         { "type": "integer", "format": "int32" },
          "featureVersion":    { "type": "integer", "format": "int32" },
          "createdSubStageId": { "type": "integer", "format": "int32", "nullable": true },
          "taxonomy":          { "$ref": "#/components/schemas/FeatureTaxonomy" }
        }
      },
      "DetachTaskBody": {
        "type": "object",
        "required": ["reassignToFeatureId"],
        "properties": {
          "reassignToFeatureId": { "type": "integer", "format": "int32", "minimum": 1 }
        }
      }
    }
  }
}